Get rudimentary node-dragging working
I say "rudimentary" because there's a notable drawback to the current
implementation: if you move the mouse too fast, the mouse cursor escapes
the node, which means the node doesn't see the event, which breaks the
dragging until you put the mouse-cursor back into the node.

A fix for this would be to put the node being dragged into the
application state, and then have a mouse-move handler on the SVG canvas
itself that's smart enough to handle node dragging. This will probably
be a good thing to work on next, but I'm quitting for the night.