Strip document from index route
It occurs to me that when we're doing an index, we care more about what
all IDs there are (and maybe when each map was created) than about the
contents of a particular map. When the user selects a particular map,
that's when we care about the map document itself, and at that point we
can fetch it using the fetch route. Omitting the document in the index
route will save bandwidth and potentially index load-time in the
frontend, especially for heavy users with several large documents.

In the end, probably not a huge deal either way. Probably not even worth
all the time I spent writing this commit message. :-) Oh well.