Allow users to be designated admins
We don't have any notion of migrations yet. So for now, best bet is to
drop the database that has the old schema installed, create a fresh
database, and restart your server to install the schema from this
commit. This will, of course, wipe the existing data, which you'll have
to recreate manually if you want. Otherwise, you can add the column
manually--just be sure to get it right!