Add login form
This commit includes authentication functions on the frontend that
should make it easy to prepend a login form (and successful login)
before other actions (like saving a map) as well.

This one turned out to be pretty involved. We rely on cookies to convey
proof of our authentication from the frontend to the backend. But by
default, XmlHttpRequests (XHRs) don't convey cookies to a remote
origin (i.e. one other than the one the javascript was originally loaded
from). You have to set the `withCredentials` flag on the XHR object to
enable this.

More info here:

https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS#Requests_with_credentials

And here:

https://github.com/mbostock/d3/issues/1397