Add delete functionality
This shows up as a link in the popup list of maps when the
third-from-left button in the toolbox is clicked. The trashcan icon
clears the canvas but does NOT delete the map from the backend, and the
message associated with that action has been updated to specify that.

Note: just like admins cannot update a map not their own, they also
cannot delete a map not their own.

If either of these things is not what's desired, just let me know.