Add rudimentary toggle-visibility wizard area
The wizard area (200px tall at the top of the page) is shown by default.
To hide it, call `hideWizard()` in the Javascript console. To show it
again, call `showWizard()`.

This is (obviously) only a rudimentary mockup, without any actual wizard
functionality, but it serves as a visual illustration for how I'm
thinking a wizard would integrate into the actual SSM app.