Fix missing column in schema
Back in commit a5c1ba1, I changed the code to assume the existence of a
`name` column. Not having a migration mechanism at the time, I simply
added the column manually to my dev database and to the production
database. This was fine, but what I should also have done was add the
column to the initial schema, so that new installations (e.g. my new
development workstation) had the column. This commit fixes the problem.