Layout nodes naively
This commit trashes all the force-layout stuff.

The simple layout implemented here is heavily skewed. The first child of
a node is directly under it (from the center), and children that are
added later are farther clockwise around the ring, which means that most
edges are going to be pointed in the clockwise direction.

A better layout would be to center the children under their parent (or
vice-versa), but this approach gets the job done for now. I may come
back and try to improve the layout later, but it's not a high
priority at the moment (as far as I know).