Resolve conflicts among dependencies
The conflicts were reported by `lein deps :tree`.

This was actually necessary to resolve some odd and blocking errors
about not being able to find clojure.tools.reader on the class path
whenever I tried to start `lein repl` in the backend project. Weird
stuff.

This post by Yogthos [1] included the verbatim error and described the
cause for him, but his solution (stop depending on the timbre logging
library) doesn't work for us here, because we're not (even indirectly)
depending on timbre. Presumably some other depended-upon library is
doing AOT like timbre, but it's not easy to find out which one, and this
commit avoids the problem anyway.

[1] http://yogthos.net/posts/2015-12-26-AOTGotchas.html