Implement "Remove" button in entry list
This entailed a fairly thorough rewrite of the entry list code. Now it
is much more idiomatic d3, making use of the so-called "general update
pattern". There are a couple of ugly spots, but overall, it's much
cleaner than it would have been if I had tried to implement the removal
functionality in the original paradigm.