Suggest completions for text entry
This is currently using mock data, but once we get real data, it will be
easy to swap in. Also, although there is currently the same list of
completions for each type of node, internally, there is a separate list
for each type. So it will be easy to separate the lists of completions
whenever we want to.

I haven't yet added the completions to the role type, since that's
significantly different. I also haven't added completions to the
resources type, since that whole step needs a lot of work.