Enable user to control label for resource
A new select box lets the user pick between using the resource's general
name (i.e. the "type of resource") for the label of the node in the map
and using the resource's specific name. It defaults to using the general
name, same as before. If the user selects name, but no name has been
provided, that's an error and the user is blocked from proceeding (with
an error message indicating what's wrong).

I also cleaned up some logic for the error message display, which was
leaving the messages displayed even after the errors had been resolved.